requirements.txt
# ─── CARVanta Dependencies ───────────────────────────────────────────────────────
# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Frontend
streamlit>=1.29.0

# Data & ML
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Database (Phase 2)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Config & Security
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# PDF Reports
reportlab>=4.0.0

# Dev & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
pip-audit>=2.6.0
